feat(connector-sdk): migrate core contract types to SDK
Move Connector, FetchContext, PageResult, AuthStatus, CapturedItem,
SyncState, SyncError, SyncErrorCode to @spool/connector-sdk. @spool/core
re-exports them for backwards compatibility and keeps a core-internal
Effect-tagged SyncError for use in Effect's typed error channel.

SDK has zero runtime dependencies.

Part of Stage D (SDK split).
